% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{add_lotek_status}
\alias{add_lotek_status}
\title{Add Lotek Status}
\usage{
add_lotek_status(
  lotek_positions,
  position_type = c("Litetrack (Standard fix)", "Litetrack (Swift fix)", "Iridium Track")
)
}
\arguments{
\item{lotek_positions}{A data frame of GPS data matching the format returned
by \code{\link{fetch_lotek_positions}}.}

\item{position_type}{One of 'Litetrack (Standard fix)',
'Litetrack (Swift fix)', or 'Iridium Track'. Determines how the information
in the FixType and RxStatus columns is parsed. See API User Manual for
details. If you have more than one position type in the same account
separate them before passing into this function (see aexamples).}
}
\value{
A tibble containing the original position data with additional
  columns containing fix status information and in some cases the number of
  satellites.
}
\description{
Parses coded values in RxStatus column of API response.
}
\section{Notes}{


  See Lotek Web Service API User Manual Revision 03 (30 MAY 2024)
}

\examples{
\dontrun{

lotek_login("demo", "PASSWORD09")

# download and parse all data - standard fixes
fixes <- fetch_lotek_positions() |>
  add_lotek_status()

# download and parse all data - swift fixes
fixes <- fetch_lotek_positions() |>
  add_lotek_status("Litetrack (Swift fix)")

# download and parse all data - iridium
fixes <- fetch_lotek_positions() |>
  add_lotek_status("Iridium Track")

# if data contains collars with different position types...
standard_dev <- c(34023, 42492)
swift_dev <- 32763

# download all, split, add status, and recombine...
fixes <- fetch_lotek_positions()
standard_fixes <- fixes |>
  dplyr::filter(DeviceID \%in\% standard_dev) |>
  add_lotek_status()
swift_fixes <- fixes |>
  dplyr::filter(DeviceID \%in\% swift_dev) |>
  add_lotek_status("Litetrack (Swift fix)")
fixes <- standard_fixes |>
  dplyr::bind_rows(swift_fixes)

# or download separately, add status, and combine
standard_fixes <- fetch_lotek_positions(device_id = standard_dev) |>
  add_lotek_status()
swift_fixes <- fetch_lotek_positions(device_id = swift_dev) |>
  add_lotek_status("Litetrack (Swift fix)")
fixes <- standard_fixes |>
  dplyr::bind_rows(swift_fixes)

}
}
\seealso{
\code{\link{fetch_lotek_positions}}
}
