% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{ats_get}
\alias{ats_get}
\title{GET}
\usage{
ats_get(path, task = "download data")
}
\arguments{
\item{path}{Character or list for the request path.}

\item{task}{Character describing the purpose of the current request.
If the request fails the message 'Failed to [task]' is displayed.}
}
\value{
HTTP response object
}
\description{
Submit an http GET request to the ATS website
}
\examples{
\dontrun{

ats_login("mary", ".")

ats_get(
  path = list(
    "download_all_events",
    "download_all_events.aspx?dw=all"
  ),
  task = "download event data"
) \%>\%
  ats_parse_txt()

}

}
\seealso{
\code{\link[httr]{GET}}
}
\keyword{internal}
