% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{ats_join_trans}
\alias{ats_join_trans}
\title{Join Fixes and Transmissions}
\usage{
ats_join_trans(pos, trans)
}
\arguments{
\item{pos}{Tibble of position data.}

\item{trans}{Tibble of corresponding transmission data.}
}
\value{
A tibble with 25 columns (see \code{\link{ats_fetch_positions}})
}
\description{
Add transmission info to fixes to apply offset and infer
  data substitutions
}
\examples{
\dontrun{

ats_login("mary", ".")

trans <- fetch_ats_transmissions()

ats_get(
  path = list(
    "download_all_data",
    paste0("Download_all_data.aspx?dw=all")
  ),
  task = "download position data"
) \%>\%
  httr::content("text", encoding = "UTF-8") \%>\%
  readr::read_csv(col_types = "ciiiiiidddiic_") \%>\%
  dplyr::rename(JulianDay = Julianday) \%>\%
  ats_join_trans(trans)

}

}
\seealso{
\code{\link{ats_fetch_positions}}
}
\keyword{internal}
