% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{ats_parse_pos}
\alias{ats_parse_pos}
\title{Parse Fixes}
\usage{
ats_parse_pos(resp, trans)
}
\arguments{
\item{resp}{HTTP response object}

\item{trans}{Tibble of corresponding transmission data}
}
\value{
A tibble with 25 columns (see \code{\link{fetch_ats_positions}})
}
\description{
Convert http response to tibble
}
\section{Notes}{


  Data can be retrieved either via \code{httr::GET} (for all
  transmissions) or \code{httr::POST} (by collar), and each request
  type can return two different formats depending on whether there is
  data available. This function handles the four different possibilities
  and returns a tibble with the same column names and types no matter
  what. GmtOffset is multiplied by -1 to correct for ATS using
  non-standard syntax. Dates in UTC and current system time zone are
  added. Column substitutions in neolink modes are reversed.
}

\examples{
\dontrun{

ats_login("mary", ".")

trans <- fetch_ats_transmissions()

ats_get(
  path = list(
    "download_all_data",
    paste0("Download_all_data.aspx?dw=all")
  ),
  task = "download position data"
) \%>\%
  ats_parse_pos(trans)

}

}
\seealso{
\code{\link{fetch_ats_positions}}
}
\keyword{internal}
