% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{ats_parse_trans}
\alias{ats_parse_trans}
\title{Parse Transmissions}
\usage{
ats_parse_trans(resp)
}
\arguments{
\item{resp}{HTTP response object}
}
\value{
A tibble with 19 columns (see \code{\link{ats_fetch_positions}})
}
\description{
Convert http response to tibble
}
\section{Notes}{


  Data can be retrieved either via \code{httr::GET} (for all
  transmissions) or \code{httr::POST} (by collar), and each request
  type can return two different formats depending on whether there is
  data available. This function handles the four different possibilities
  and returns a tibble with the same column names and types no matter
  what. GmtOffset is multiplied by -1 to correct for ATS using
  non-standard syntax. Dates in UTC and current system time zone are
  added. Neolink events are split into separate columns.
}

\examples{
\dontrun{

ats_login("mary", ".")

trans <- fetch_ats_transmissions()

ats_get(
  path = list(
    "download_all_transmission",
    "download_all_transmission.aspx?dw=all"
  ),
  task = "download transmission data"
) \%>\%
  ats_parse_trans()

}

}
\seealso{
\code{\link{ats_fetch_transmissions}}
}
\keyword{internal}
