% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{ats_parse_txt}
\alias{ats_parse_txt}
\title{Parse Text}
\usage{
ats_parse_txt(resp, ...)
}
\arguments{
\item{resp}{HTTP response object}

\item{...}{Additional arguments passed to \code{readr::read_csv}}
}
\value{
A tibble
}
\description{
Convert raw text file in HTTP response to tibble
}
\examples{
\dontrun{

ats_login("mary", ".")

ats_get(
  path = list(
    "download_all_events",
    "download_all_events.aspx?dw=all"
  ),
  task = "download event data"
) \%>\%
  ats_parse_txt()

}

}
\keyword{internal}
