% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{ats_parse_xml}
\alias{ats_parse_xml}
\title{Parse XML}
\usage{
ats_parse_xml(resp)
}
\arguments{
\item{resp}{HTTP response object}
}
\value{
A tibble
}
\description{
Convert raw xml in HTTP response to tibble
}
\examples{
\dontrun{

ats_login("mary", ".")

ats_post(
  path = "Servidor.ashx",
  body = list(
    consulta = "download_txt_collars",
    type = "004",
    parameter1 = "",
    parameter2 = ""
  ),
  task = "download position data"
) \%>\%
  ats_parse_xml()

}

}
\keyword{internal}
