% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{ats_post}
\alias{ats_post}
\title{POST}
\usage{
ats_post(path, body = list(), task = "download data")
}
\arguments{
\item{path}{Character or list for the request path}

\item{body}{Named list of query parameters}

\item{task}{Character describing the purpose of the current request.
If the request fails the message 'Failed to [task]' is displayed.}
}
\value{
Response object
}
\description{
Submit an http POST request to the ATS website
}
\examples{
\dontrun{

ats_login("mary", ".")

ats_select_collars("044286")

ats_post(
  path = "Servidor.ashx",
  body = list(
    consulta = "download_trans_collars"
  ),
  task = "download transmission data"
) \%>\%
  ats_parse_xml()

}

}
\seealso{
\code{\link{httr::POST}}
}
\keyword{internal}
