% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{ats_select_collars}
\alias{ats_select_collars}
\title{Select Collars}
\usage{
ats_select_collars(device_id)
}
\arguments{
\item{device_id}{A single device id, or a list or vector of device ids.}
}
\value{
Logical, TRUE if successful
}
\description{
Submit an http POST request to create a cookie of selected
  collars for subsequent requests
}
\examples{
\dontrun{

ats_login("mary", ".")

all_collars <- fetch_ats_devices()
some_collars <- sample(all_collars, 5)
ats_select_collars(some_collars)

}

}
\keyword{internal}
