% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{ats_trans_dates}
\alias{ats_trans_dates}
\title{Transmission Dates}
\usage{
ats_trans_dates(trans)
}
\arguments{
\item{trans}{A tibble of transmission data}
}
\value{
A tibble with dates in UTC and local (system) time added
}
\description{
Add date columns to transmission data
}
\examples{
\dontrun{

ats_login("mary", ".")

ats_select_collars("044286")

ats_post(
  path = "Servidor.ashx",
  body = list(
    consulta = "download_trans_collars"
  ),
  task = "download transmission data"
) \%>\%
  ats_parse_xml() \%>\%
  dplyr::mutate(
    DateCT = as.POSIXct(
      fecha,
      tz = "America/Menominee"
    ),
    GmtOffset = as.numeric(gmt) * -1
  ) \%>\%
  ats_trans_dates()

}

}
\keyword{internal}
