% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ats_auth.R
\name{check_cookie}
\alias{check_cookie}
\title{Check Cookie}
\usage{
check_cookie(url, cookie)
}
\arguments{
\item{url}{http hostname (base url)}

\item{cookie}{name of the cookie to check}
}
\value{
True if cookie exists, false if not
}
\description{
Check that a particular cookie exists for a certain site
}
\examples{
\dontrun{

check_cookie(ats_base_url, "ARRAffinity")

}

}
\keyword{internal}
