% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ats_auth.R
\name{clear_cookie}
\alias{clear_cookie}
\title{Remove Cookie}
\usage{
clear_cookie(url, cookie)
}
\arguments{
\item{url}{http hostname (base url)}

\item{cookie}{name of the cookie to remove}
}
\value{
named character vector for \code{httr::set_cookies}
}
\description{
Remove a certain cookie from a request handle
}
\examples{
\dontrun{

httr::GET(
  url = ats_base_url,
  path = list(
  "download_all_transmission",
  "download_all_transmission.aspx?dw=new"
  ),
  httr::set_cookies(clear_cookie(ats_base_url, "cgca"))
)

}

}
\keyword{internal}
