% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllr_add_id.R
\name{cllr_add_id}
\alias{cllr_add_id}
\title{Add id column}
\usage{
cllr_add_id(x, id_vals)
}
\arguments{
\item{x}{object inherting from data.frame}

\item{id_vals}{values (character) to include in a newly created id column, these are the unique identifier for each individual}
}
\value{
data.frame
}
\description{
If your data does not have unique identifiers for each animal cllr_add_id will create that column and insert values you provide.
}
\details{
When adding a new column to data using cllr_add_id the user should supply either a length 1 character string or a character vector equal in length to the number of observations in the data. The latter is not recommended.
}
\examples{
# Add unique identifiers
df <- data.frame(val = 1:3)
cllr_add_id(df, "1A")
}
\seealso{
cllr_rename_id
}
