% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_scale_color.R
\name{cllr_colors}
\alias{cllr_colors}
\title{Grab any combination of colors for our palettes}
\usage{
cllr_colors(...)
}
\arguments{
\item{...}{Additional arguments which can be passed to ggplot2::disrete_scale()
or ggplot2::scale_color_gradientn(), used when discrete is
true or false, respectively}
}
\description{
Grab any combination of colors for our palettes
}
