% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllr_dl_vectronics.R
\name{cllr_dl_vectronics}
\alias{cllr_dl_vectronics}
\title{Download GPS collar data from Vectronics web server}
\usage{
cllr_dl_vectronics(key_dir = key_dir, base_url = NULL,
  after_data_id = NULL, type = "gps", start_date = NULL)
}
\arguments{
\item{key_dir}{A length 1 character string describing the path to the directory (on your machine) that houses key files.}

\item{base_url}{NULL or the url to build from, the default NULL should
suffice in almost all cases}

\item{after_data_id}{All data types have a unique ID maintained by the
manufacturer.  Use this parameter to download data after the supplied data
ID.  Must be equal in length to collar_id and collar_key.  Only one of
start_date and after_data_id may be supplied.}

\item{type}{The data type, options include "gps", "act", "mit", "mor", "prx",
"SEP" and "TRAP".  See details.}

\item{start_date}{A character vector specifying date and time as
DD-MM-YYYYTHH:MM:SS (the T is required, for example 01-01-2018T00:00:00).
Supplying this parameter will cause the API to only return data collected
after the supplied date.  Must be equal in length to collar_id and
collar_key.  Only one of start_date and after_data_id may be supplied.}
}
\value{
A single tibble of collar data from every collar id.
}
\description{
Download Vectronics GPS collar data for each collar id or "key file" in the directory "key_dir".
}
\examples{
cllr_dl_vectronics(key_dir = "./keys")
}
