% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllr_fetch_csv.R
\name{cllr_fetch_csv}
\alias{cllr_fetch_csv}
\title{Read collar data stored in CSV file(s)}
\usage{
cllr_fetch_csv(file_path, skip = 0, ...)
}
\arguments{
\item{file_path}{The full path to one or more csv files to read}

\item{skip}{Scalar integer, the number of rows to skip at the top of the
file, see details}

\item{...}{Other parameters to pass to readr::read_csv}
}
\value{
a tibble with columns collar_id, date_time, x, y plus any additional
columns requested via aux_cols
}
\description{
Read collar data stored in CSV file(s)
}
\examples{
#  Define file path
fpath <- system.file(
  "extdata",
  "lotek_2.csv",
  package = "collar",
  mustWork = TRUE
)

#  Read file
ltk <- readr::read_csv(fpath)

#  Example with large header
fpath <- system.file(
"extdata",
"telonics.csv",
package = "collar",
mustWork = TRUE
)

#  Read in whole data, use it to find first row in next step
r1 <- readr::read_csv(fpath)

#  Read in Telonics data skipping header
tlncs <- readr::read_csv(fpath, skip = which(r1[,1] == "Acquisition Time"))

}
