% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_gps.R
\name{cllr_morph_gps}
\alias{cllr_morph_gps}
\title{Standardize GPS data to collar format}
\usage{
cllr_morph_gps(x = NULL, id_col = NULL, dt_col = NULL,
  dt_format = "\%d/\%m/\%Y \%H:\%M:\%S", lon_col = NULL,
  lat_col = NULL, meta = list(), extra = character())
}
\arguments{
\item{x}{data.frame of GPS data with at least columns ID, Latitude, Longitude and Date-Time}

\item{id_col}{Unquoted name of the ID column that signals the unique identifier for this collar}

\item{lon_col}{Unquoted name of the column containing Longitude location information.  See Details.}

\item{lat_col}{Unquoted name of the column containing Latitude location information.  See Details.}

\item{meta}{Key-value pairs supplied as a named list specifying new columns and values to be added to data.  See Details.}

\item{extra}{Unquoted column names of those columns beyond id, datetime, lon and lat that you wish to retain in the data set.}
}
\description{
Standardize GPS data to collar format
}
\details{
stuff
}
