% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_scale_color.R
\name{cllr_pal}
\alias{cllr_pal}
\title{Access Palettes}
\usage{
cllr_pal(palette = "main", reverse = F, ...)
}
\arguments{
\item{palette}{Character name of palette in cllr_palettes}

\item{reverse}{Boolean; Should the colors be reversed?}

\item{...}{Additional arguments which can be passed to ggplot2::disrete_scale()
or ggplot2::scale_color_gradientn(), used when discrete is
true or false, respectively}
}
\description{
Access any of the color palettes available
}
