% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_scale_color.R, R/style_scale_fill.R
\docType{data}
\name{cllr_palettes}
\alias{cllr_palettes}
\title{Create palettes}
\format{An object of class \code{list} of length 4.}
\usage{
cllr_palettes

cllr_palettes
}
\description{
The palettes we've decided to create, can be moved.

The palettes we've decided to create, can be moved.
}
\keyword{datasets}
