% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllr_remove_header.R
\name{cllr_remove_header}
\alias{cllr_remove_header}
\title{A function to find and remove data headers form collar data based on the latitude column}
\usage{
cllr_remove_header(x, lat_col, rm_header)
}
\arguments{
\item{x}{A dataframe containing collar data}

\item{lat_col}{The unqouted name of the latitude or any other column in the data frame}

\item{rm_header}{Logical indicating whether to remove header completely}
}
\value{
tibble
}
\description{
A function to find and remove data headers form collar data based on the latitude column
}
\examples{
ugly_file <- read.csv(system.file("extdata", "telonics.csv", package = "collar"))
# Use backticks for column names with spaces
collar::cllr_remove_header(ugly_file, `GPS Latitude`, rm_header = TRUE)
}
