% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collar.R
\docType{package}
\name{collar-package}
\alias{collar}
\alias{collar-package}
\title{collar Utilities to help people get their data from GPS collars or other sensors}
\description{
Modern GPS collars send data to end users via clunky websites, email accounts or proprietary applications. The purpose of this package is to automate the process of accessing your data.

To learn more about collar, start with the vignettes:
coming soon=(
}
\section{Package options}{

\describe{
\item{`none`}{There are no package specific options at this time}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Huh/collar}
  \item Report bugs at \url{https://github.com/Huh/collar/issues}
}

}
\author{
\strong{Maintainer}: Josh Nowak \email{josh.nowak@speedgoat.io}

Authors:
\itemize{
  \item Forest Hayes \email{forest.hayes@umontana.edu}
  \item T.J. Clark \email{tjclark1873@gmail.com}
  \item Hans Martin \email{hmartin289@gmail.com}
  \item Eric Newkirk \email{eric.newkirk@speedgoat.io}
}

Other contributors:
\itemize{
  \item SpeedGoat Wildlife Solutions [copyright holder, funder]
}

}
