% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_scale_color.R, R/style_scale_fill.R
\docType{data}
\name{colors}
\alias{colors}
\title{Colors}
\format{An object of class \code{character} of length 12.}
\usage{
colors

colors
}
\description{
# this is from: coolors.co

# this is from: coolors.co
}
\keyword{datasets}
