% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{fetch_ats}
\alias{fetch_ats}
\title{Scrape data from ATS website}
\usage{
fetch_ats(
  bttn_nm = "ctl00$ContentPlaceHolder1$DownloadAll3",
  usr,
  pwd,
  base_url = "atsidaq.net",
  rename_fun = adj_col_nms
)
}
\arguments{
\item{bttn_nm}{The name of the button to download data from}

\item{usr}{The username associated with the account}

\item{pwd}{The password used with the supplied username}

\item{base_url}{The URL where the function attempts to gain access. By default this is set to "atsidaq.net" which is the domain used for recent GPS collars. Use of other domains (e.g. "atsdat.net") is supported by specifying a URL value in order to support older Iridium based collars and for future compatibility.}

\item{rename_fun}{A function used to rename columns, defaults to adj_col_nms, see details}
}
\value{
A tibble
}
\description{
Scrape data from ATS website
}
\details{
The url where the function attempts to gain access is atsidaq.net.

To download from a new button name use the developer tools in your browser and record the name of the button as described there. For example, the download all button has name 'ctl00$ContentPlaceHolder1$DownloadAll3'. See the vignette for more details about finding button names.

Column names are adjusted using a custom function, but the user can pass any function they want to manipulate column names (e.g. make.names). The default removes non-ASCII characters, coerces all characters to lower case and replaces "." with "_".
}
\examples{
\dontrun{
  fetch_ats(usr = "my_username", pwd = "secret_code")
}
}
