% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{fetch_ats_config}
\alias{fetch_ats_config}
\title{Download Collar Configuration Data from ATS Website}
\usage{
fetch_ats_config()
}
\value{
A tibble with 6 columns:
\describe{
  \item{CollarSerialNumber}{ATS Collar ID (character)}
  \item{Email}{Email recipient for alerts (character)}
  \item{Active}{Is collar currently active (logical)}
  \item{Phone Num SMS}{SMS recipient for alerts (character)}
  \item{FTP Url}{URL for retrieving data via ftp (character)}
  \item{RestEndPoint}{Is data available via REST API (logical)}
}
}
\description{
Retrieves configuration information for all collars in the
  current account
}
\examples{
\dontrun{

ats_login("mary", ".")

# get configuration details for all collars in this account
collar_details <- fetch_ats_config()

ats_logout()

}

}
\seealso{
\code{\link{ats_login}} for logging into an ATS account,
  \code{\link{fetch_ats_positions}} for downloading GPS data,
  \code{\link{fetch_ats_transmissions}} for downloading transmission
  data, \code{\link{fetch_ats_events}} for downloading alerts, and
  \code{link{fetch_ats_devices}} for downloading a list of available
  collars
}
