% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{fetch_ats_devices}
\alias{fetch_ats_devices}
\title{Download a List of Devices from ATS Website}
\usage{
fetch_ats_devices(filter = "all")
}
\arguments{
\item{filter}{A single character value for filtering the results by
status. If an invalid filter value is provided
\code{fetch_ats_devices} returns a list of all devices with a warning.
Valid filter values include:
  \itemize{
    \item{All (Default) - a list of all collars}
    \item{Active - Only active collars}
    \item{Inactive - Only inactive collars}
    \item{Low_batt - Active collars with low battery alerts}
    \item{Mort - Active collars with mortality alerts}
    \item{Birth - Active collars with birth event alerts}
  }}
}
\value{
A character vector of device ids
}
\description{
Retrieves a list of devices (collars), optionally
  filtered by status
}
\examples{
\dontrun{

ats_login("mary", ".")

# get ids for all collars in this account
collar_list <- fetch_ats_devices()

# get ids for collars active collars
collar_list <- fetch_ats_devices("Active")

# get ids for collars inactive collars
collar_list <- fetch_ats_devices("Inactive")

# get ids for collars with low battery
collar_list <- fetch_ats_devices("Low_batt")

# get ids for collars in mortality
collar_list <- fetch_ats_devices("Mort")

# get ids for collars with birth events triggered
collar_list <- fetch_ats_devices("Birth")

ats_logout()

}

}
\seealso{
\code{\link{ats_login}} for logging into an ATS account,
  \code{\link{fetch_ats_config}} for downloading collar configurations,
  \code{\link{fetch_ats_positions}} for downloading GPS data,
  \code{\link{fetch_ats_transmissions}} for downloading transmission
  data, and \code{\link{fetch_ats_events}} for downloading alerts
}
