% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{fetch_ats_events}
\alias{fetch_ats_events}
\title{Download Event Data from ATS Website}
\usage{
fetch_ats_events()
}
\value{
A tibble with 8 columns:
\describe{
  \item{CollarSerialNumber}{ATS Collar ID (character)}
  \item{DateCT}{Timestamp from server (US Cental time, POSIXct))}
  \item{DateUTC}{Timestamp in UTC/GMT (POSIXct)}
  \item{DateLocal}{Timestamp in current system time zone (POSIXct)}
  \item{Birth}{VIT birth event (character)}
  \item{Fawn0}{Main collar event or neolink slot 0 event (character)}
  \item{Fawn1}{Neolink slot 1 event (character)}
  \item{Fawn2}{Neolink slot 2 event (character)}
}
}
\description{
Retrieves all undownloaded events (a.k.a. alerts)
}
\examples{
\dontrun{

ats_login("mary", ".")

# get undownloaded events for all collars in this account
alerts <- fetch_ats_events()

ats_logout()

}

}
\seealso{
\code{\link{ats_login}} for logging into an ATS account,
  \code{\link{fetch_ats_config}} for downloading collar configurations,
  \code{\link{fetch_ats_positions}} for downloading GPS data,
  \code{\link{fetch_ats_transmissions}} for downloading transmission
  data, and \code{link{fetch_ats_devices}} for downloading a list of
  available collars
}
