% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{fetch_ats_positions}
\alias{fetch_ats_positions}
\title{Download GPS Fixes from ATS Website}
\usage{
fetch_ats_positions(
  device_id = NULL,
  start = NULL,
  end = NULL,
  n = NULL,
  new = FALSE
)
}
\arguments{
\item{device_id}{A single device id, or a list or vector of device ids,
or NULL for all devices associated with current account. Overrides
the new parameter when specified.}

\item{start, end}{Currently ignored (see Notes).}

\item{n}{A single integer specifying how many fixes to return per
collar (sorted by recency). Valid values are 5 and 10.}

\item{new}{A logical flag. When new = true only data that hasn't been
previously downloaded is returned. If device_id is specified the flag
is ignored.}
}
\value{
A tibble with 25 columns:
\describe{
  \item{CollarSerialNumber}{ATS Collar ID (character)},
  \item{Year}{Two digit year (integer)},
  \item{JulianDay}{Julian day (1:366, integer)},
  \item{Hour}{Hour (integer)},
  \item{Minute}{Minute (integer)},
  \item{Activity}{Activity sensor reading (integer)},
  \item{Temperature}{Ambient temperature (Celsius, integer)},
  \item{Latitude}{Latitude (decimal degrees, numeric)},
  \item{Longitude}{Longitude (decimal degrees, numeric)},
  \item{HDOP}{Horizontal Dilution of Precision (numeric)},
  \item{NumSats}{Number of GPS satellites (integer)},
  \item{FixTime}{GPS Fix Time (integer)},
  \item{`2D/3D`}{GPS fix dimension (integer)},
  \item{DateOffset}{Timestamp from raw data with fixed offset applied
    (POSIXct)},
  \item{GmtOffset}{Offset from GMT in hours, mutliplied by negative one
    to correct for non-standard ATS usage (numeric)},
  \item{DateUTC}{Timestamp in UTC/GMT (POSIXct)},
  \item{DateLocal}{Timestamp in current system time zone (POSIXct)},
  \item{VITTemp}{VIT temperature (Celsius, integer)},
  \item{VITLight}{VIT light sensor reading (integer) (integer)},
  \item{VITComm}{VIT communication count (integer)},
  \item{Fawn0Comm}{Fawn slot 0 communication count (integer)},
  \item{Fawn1Comm}{Fawn slot 1 communication count (integer)},
  \item{Fawn2Comm}{Fawn slot 2 communication count (numeric)},
  \item{TransDateUTC}{Transmission timestamp in UTC/GMT (POSIXct)},
  \item{TransDateLocal}{Transmission timestamp in current system time
    zone (POSIXct)}
}
}
\description{
Retrieves GPS data optionally filtered by date, last n
  positions, or collar
}
\section{Notes}{


  Currently filtering by date is broken on the ATS site - using the
  date range filters when downloading data returns an internal server
  error. The functionality is included here in the hopes that it will
  be fixed in the near future, but currently date filters are ignored.
}

\examples{
\dontrun{

ats_login("mary", ".")

# get undownloaded fixes for all collars in this account
fixes <- fetch_ats_positions(new = TRUE)

# get all fixes for all collars in this account
fixes <- fetch_ats_positions()

# get all fixes for specific collars
collar_list <- sample(fetch_ats_devices(), 10)
fixes <- fetch_ats_positions(collar_list)

# get all fixes for collars in mortality
collar_list <- fetch_ats_devices("mortality")
fixes <- fetch_ats_positions(device_id = collar_list)

# get last 10 fixes for certain collars
fixes <- fetch_ats_positions(device_id = collar_list, n = 10)

# get fixes in 2019 for all collars
fixes <- fetch_ats_positions(
  start = as.POSIXct("2019-01-01"),
  end = as.POSIXct("2020-01-01")
)

# get fixes in 2019 for certain collars
fixes <- fetch_ats_positions(
  device_id = collar_list,
  start = as.POSIXct("2019-01-01"),
  end = as.POSIXct("2020-01-01")
)

# get undownloaded fixes for a single collar
fixes <- fetch_ats_positions("044286", new = TRUE)

ats_logout()

}

}
\seealso{
\code{\link{ats_login}} for logging into an ATS account,
  \code{\link{fetch_ats_config}} for downloading collar configurations,
  \code{\link{fetch_ats_transmissions}} for downloading transmission
  data, \code{\link{fetch_ats_events}} for downloading alerts, and
  \code{link{fetch_ats_devices}} for downloading a list of available
  collars
}
