% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ats.R
\name{fetch_ats_transmissions}
\alias{fetch_ats_transmissions}
\title{Download Transmission Data from ATS Website}
\usage{
fetch_ats_transmissions(device_id = NULL, new = FALSE)
}
\arguments{
\item{device_id}{A single device id, or a list or vector of device ids,
or NULL for all devices associated with current account. Overrides
the new parameter when specified. If the device ids returned by
\code{fetch_ats_devices} have leading zeros they should be included
in the \code{device_id} parameter as well
(e.g. \code{device_id = "012345")}).}

\item{new}{A logical flag. When new = true only data that hasn't been
previously downloaded is returned. If device_id is specified the flag
is ignored.}
}
\value{
A tibble with 20 columns:
\describe{
  \item{CollarSerialNumber}{ATS Collar ID (character)}
  \item{DateCT}{Timestamp from server (US Cental time, POSIXct))}
  \item{DateUTC}{Timestamp in UTC/GMT (POSIXct)}
  \item{DateLocal}{Timestamp in current system time zone (POSIXct)}
  \item{NumberFixes}{Number of fixes transmitted (integer)}
  \item{BattVoltage}{Battery voltage (numeric)}
  \item{Mortality}{Mortality message (character)}
  \item{BreakOff}{Breakoff message (character)}
  \item{GpsOnTime}{GPS module on time (integer)}
  \item{SatOnTime}{Satellite module on time (integer)}
  \item{SatErrors}{Number of satellite errors (integer)}
  \item{GmtOffset}{Offset (in hours) from UTC/GMT (numeric)}
  \item{LowBatt}{Is collar in low battery mode (logical)}
  \item{Birth}{VIT birth event (character)}
  \item{Fawn0}{Neolink slot 0 event (character)}
  \item{Fawn1}{Neolink slot 1 event (character)}
  \item{Fawn2}{Neolink slot 2 event (character)}
  \item{Latitude}{Latitude in decimal degrees (numeric)}
  \item{Longitude}{Latitude in decimal degrees (numeric)}
  \item{CEPradius_km}{Circular Error Probability in km (integer)}
}
}
\description{
Retrieves all transmissions or undownloaded transmissions,
  optionally filtered by collar
}
\examples{
\dontrun{

ats_login("mary", ".")

# get undownloaded transmissions for all collars in this account
trans <- fetch_ats_transmissions(new = TRUE)

# get all transmissions for all collars in this account
trans <- fetch_ats_transmissions()

# get all transmissions for specific collars
collar_list <- sample(fetch_ats_devices(), 10)
trans <- fetch_ats_transmissions(device_id = collar_list)

ats_logout()

}

}
\seealso{
\code{\link{ats_login}} for logging into an ATS account,
  \code{\link{fetch_ats_config}} for downloading collar configurations,
  \code{\link{fetch_ats_positions}} for downloading GPS data,
  \code{\link{fetch_ats_events}} for downloading alerts, and
  \code{link{fetch_ats_devices}} for downloading a list of available
  collars
}
