% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{fetch_lotek_alerts}
\alias{fetch_lotek_alerts}
\title{Download Alerts via Lotek API}
\usage{
fetch_lotek_alerts()
}
\value{
A tibble with information about mortalities, births, etc.
}
\description{
Retrieves alerts for all devices associated with the account
  currently logged in via lotek_login.
}
\examples{
\dontrun{
lotek_login("demo", "PASSWORD09")

# get all alerts for this account
alerts <- fetch_lotek_alerts()
}
}
\seealso{
\code{\link{fetch_lotek_devices}} for downloading a list of
  available devices, and \code{\link{fetch_lotek_positions}} for downloading
  GPS data.
}
