% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{fetch_lotek_devices}
\alias{fetch_lotek_devices}
\title{Download Device List via Lotek API}
\usage{
fetch_lotek_devices()
}
\value{
A tibble with device information.
}
\description{
Retrieves a list of devices (collars) associated
  with the account currently logged in via lotek_login.
}
\examples{
\dontrun{
lotek_login("demo", "PASSWORD09")

# get list of collars for this account
collars <- fetch_lotek_devices()

# get fixes for the first collar
fixes <- fetch_lotek_positions(device_id = collars[[1, "DeviceID"]])
}
}
\seealso{
\code{\link{fetch_lotek_alerts}} for downloading alerts such as
  mortality events, and \code{\link{fetch_lotek_positions}} for downloading
  GPS data.
}
