% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_trktg.R
\name{fetch_trktg_devices}
\alias{fetch_trktg_devices}
\title{Download Device List from track tag web portal}
\usage{
fetch_trktg_devices()
}
\value{
A tibble with 3 character columns:
\describe{
  \item{api_id}{Six digit internal id used in api calls},
  \item{trktg_id}{Nine digit transmitter id (serial number)},
  \item{user_id}{User label assigned through track tag portal}
}
}
\description{
Retrieves a list of devices (transmitters) associated
  with the account currently logged in via \code{trktg_login}.
}
\examples{
\dontrun{
trktg_login("some_user", "some_users_pw")

tt_devices <- fetch_trktg_devices()

tt_fixes <- fetch_trktg_positions(device_id = tt_devices$user_id[1])

trktg_logout()
}
}
\seealso{
\code{\link{fetch_trktg_positions}} for downloading GPS data.
}
