% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Fake function to show structure of comments and code}
\usage{
foo(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric of length x
}
\description{
Fake function to show structure of comments and code
}
\examples{
foo(4)
}
