% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllr_dl_vectronics.R
\name{get_id_from_fnames}
\alias{get_id_from_fnames}
\alias{get_id_from_key}
\title{Extract the collar ID from either key filenames or the actual key XML}
\usage{
get_id_from_fnames(key_dir, ...)

get_id_from_key(key_dir)
}
\arguments{
\item{key_dir}{A length 1 character string describing the path to the directory (on your machine) that houses key files}

\item{...}{Other arguments to pass to key_paths}
}
\value{
character vector of collar IDs
}
\description{
Extract the collar ID from either key filenames or the actual key XML
}
\details{
Vectronics includes the collar ID in the key file name as well as within the XML of the .keyx file.  These functions give the user the option of either extracting the information from the name of the file(s) get_id_from_fnames or reading the ID encoded in the key file get_id__from_key.
}
\examples{
get_id_from_fnames(
  system.file("extdata", package = "CollarDownloadeR")
)
get_id_from_key(
  system.file("extdata", package = "CollarDownloadeR")
)
}
