% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vectronics.R
\name{get_id_from_key}
\alias{get_id_from_key}
\title{Extract collar ID from key files}
\usage{
get_id_from_key(key_paths)
}
\arguments{
\item{key_paths}{The full path to one or more key files}
}
\value{
character vector of collar IDs
}
\description{
Extract collar ID from key files
}
\details{
Vectronics includes the collar ID in the XML of the .keyx file.  The get_id_from_key function gives the user the ability to read the ID encoded in the key file. This function was originally intended for internal package use, but is exported to package users in the hope it is helpful for reporting or other purposes. There is no reason to call it within the established package workflow.
}
\examples{
get_id_from_key(
  get_paths(system.file("extdata", package = "collar"))
)
}
\seealso{
Other get: 
\code{\link{get_keys}()},
\code{\link{get_paths}()}
}
\concept{get}
