% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vectronics.R
\name{get_keys}
\alias{get_keys}
\title{Extract alphanumeric key from key files}
\usage{
get_keys(key_paths)
}
\arguments{
\item{key_paths}{The full path to one or more key files}
}
\value{
A character vector of keys
}
\description{
The key is needed to call the Vectronics API and given the length of the key copy and paste errors seems likely.  The functions here help users access the keys in a programmatic fashion. We don't anticipate much need for users to use this function as it was originally intended to be internal to the package, but if listing keys is needed for reporting purposes or otherwise the function is made available.
}
\examples{

my_key <- system.file(
  "extdata",
  "Collar1000001_Registration.keyx",
  package = "collar"
)

# Extract one key
get_keys(my_key)

# Extract keys from "several" files
get_keys(c(my_key, my_key, my_key))
}
\seealso{
Other get: 
\code{\link{get_id_from_key}()},
\code{\link{get_paths}()}
}
\concept{get}
