% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vectronics.R
\name{get_paths}
\alias{get_paths}
\title{Extract path(s) to files in a directory}
\usage{
get_paths(key_dir, ext = "keyx$", ...)
}
\arguments{
\item{key_dir}{Path to a directory}

\item{ext}{Defaults to keyx, which should be correct for Vectronics key files, but could also be a regular expression used to match any file extension.}

\item{...}{other arguments to pass to list.files, see ?list.files}
}
\value{
full path to each file matching \code{ext}
}
\description{
Extract path(s) to files in a directory
}
\examples{
# Example collar key
get_paths(
  system.file("extdata", package = "collar"),
  ext = "keyx$"
)

# Example trap key
get_paths(
  system.file("extdata", "TrapKeys", package = "collar")
)

# All csv files included in this package
get_paths(
  system.file("extdata", package = "collar"),
  ext = "csv$"
)

# All keyx files included with this package
get_paths(system.file("extdata", package = "collar"), recursive = TRUE)
}
\seealso{
Other get: 
\code{\link{get_id_from_key}()},
\code{\link{get_keys}()}
}
\concept{get}
