% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vectronics.R
\name{key_paths}
\alias{key_paths}
\title{Create paths to key files}
\usage{
key_paths(key_dir, ext = "keyx")
}
\arguments{
\item{key_dir}{path to a single directory containing, potentially many, key files}

\item{ext}{Defaults to keyx, which should be correct for key files, but could also be a regular expression used to match file extension.}
}
\value{
Character vector defining path to each file matching \code{ext}
}
\description{
Create paths to key files
}
\examples{
collar:::key_paths(
  system.file("extdata", package = "collar")
)
}
\keyword{internal}
