% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{lotek_auth_token}
\alias{lotek_auth_token}
\title{Format current access token}
\usage{
lotek_auth_token(lotek_auth)
}
\arguments{
\item{lotek_auth}{a list matching the output of lotek_get_token}
}
\value{
access token as string for authenticating API calls
}
\description{
Checks token is valid and returns as character for API calls.
  Note that the current token in the environment is intentionally
  overwritten in this function, in case it was refreshed in the call to
  lotek_check_auth_token
}
\keyword{internal}
