% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{lotek_check_token}
\alias{lotek_check_token}
\title{Check and when appropriate refresh lotek auth token}
\usage{
lotek_check_token(lotek_auth)
}
\arguments{
\item{lotek_auth}{a list matching the output of lotek_get_token}
}
\value{
the same list, or an updated list if tokens are refreshed
}
\description{
Check and when appropriate refresh lotek auth token
}
\keyword{internal}
