% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{lotek_get_alerts}
\alias{lotek_get_alerts}
\title{Retrieve alerts from Lotek site via API}
\usage{
lotek_get_alerts(lotek_auth)
}
\arguments{
\item{lotek_auth}{a list matching the output of lotek_get_token}
}
\value{
A tibble with alert data
}
\description{
Retrieves alert data from the lotek API.
  Retrieves alerts for all devices associated with the account
  currently logged in via lotek_auth.
}
\examples{

tkn <- lotek_get_token(user = "speedgoat", pw = "L0tek1")

ltk_alerts <- lotek_get_alerts(tkn)

}
