% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{lotek_get_device_list}
\alias{lotek_get_device_list}
\title{Retrieve device list from Lotek site via API}
\usage{
lotek_get_device_list(lotek_auth)
}
\arguments{
\item{lotek_auth}{a list matching the output of lotek_get_token}
}
\value{
A tibble with device information
}
\description{
Retrieves a list of devices (collars) associated
  with the account currently logged in via lotek_auth.
}
\examples{
\dontrun{
tkn <- lotek_get_token(user = "speedgoat", pw = "L0tek1")

ltk_collars <- lotek_get_device_list(tkn)
}
}
