% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{lotek_get_positions}
\alias{lotek_get_positions}
\title{Retrieve position (fix) data from Lotek site via API}
\usage{
lotek_get_positions(
  lotek_auth,
  device_id = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{lotek_auth}{a list matching the output of lotek_get_token}

\item{device_id}{a single device id as character, or a
single character value containing multiple device ids separated
by commas, or NULL for all devices associated with current account}

\item{start_date}{only fixes at or after start_date will be inlcluded
in the output. Must be in the format YYYY-MM-DDTHH:MM:SSZ. Don't
forget the 'T' and 'Z'.  If NULL Jan 1 1970 is used.}

\item{end_date}{only fixes at or before end_date are included,
analagous to start_date above.  If NULL current date/time is used.}
}
\value{
A tibble containing position data
}
\description{
Retrieve position (fix) data from Lotek site via API
}
\examples{

tkn <- lotek_get_token(user = "speedgoat", pw = "L0tek1")

ltk_fixes_2020 <- lotek_get_positions(tkn,
                                      start_date = "2020-01-01T00:00:00Z",
                                      end_date = "2021-01-01T00:00:00Z",)

ltk_fixes_collar123 <- lotek_get_positions(tkn,
                                           device_id = "123")

}
