% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{lotek_get_token}
\alias{lotek_get_token}
\title{Initial login to lotek web service}
\usage{
lotek_get_token(user, pw)
}
\arguments{
\item{user}{username}

\item{pw}{password}
}
\value{
a list containing the json response,
  inluding access_token, refresh_token, and .expires
  This list is reused for authenticating subsequent calls
}
\description{
Send username and password info to Lotek API
  to retrieve authentication token
}
\examples{

tkn <- lotek_get_token(user = "speedgoat", pw = "L0tek1")

}
