% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotek_auth.R
\name{lotek_login}
\alias{lotek_login}
\title{Authenticate to Lotek API}
\usage{
lotek_login(user, pw)
}
\arguments{
\item{user}{username}

\item{pw}{password}
}
\value{
True if login succeeds, false if not.
}
\description{
Send username and password info to Lotek API
  to log in to a user account.
}
\examples{

lotek_login("demo", "PASSWORD09")

alerts <- fetch_lotek_devices()

lotek_logout()

}
\seealso{
\code{\link{lotek_logout}} for closing the session.
}
