% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{lotek_num_sats}
\alias{lotek_num_sats}
\title{Parse Lotek Number of Satellites}
\usage{
lotek_num_sats(int_values)
}
\arguments{
\item{int_values}{Vector of integers from API response}
}
\value{
Intger vector representing number of satellites for each fix
}
\description{
Converts integers in RxStatus column of API response to number
  of satellites.
}
\section{Notes}{


  See Lotek Web Service API User Manual Revision 03 (30 MAY 2024)
}

\seealso{
\code{\link{lotek_status_code}}
}
\keyword{internal}
