% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{lotek_refresh_token}
\alias{lotek_refresh_token}
\title{Refresh current access token}
\usage{
lotek_refresh_token(lotek_auth)
}
\arguments{
\item{lotek_auth}{a list matching the output of lotek_get_token}
}
\value{
a list matching the input with new values
}
\description{
Retrieves new authentication token from Lotek API
}
\keyword{internal}
