% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lotek.R
\name{lotek_status_code}
\alias{lotek_status_code}
\title{Parse Lotek Status Code}
\usage{
lotek_status_code(int_values)
}
\arguments{
\item{int_values}{Vector of integers from API response}
}
\value{
Vector of integer status codes
}
\description{
Converts integers in RxStatus column of API response to integer
  fix status codes.
}
\section{Notes}{


  See Lotek Web Service API User Manual Revision 03 (30 MAY 2024)
}

\seealso{
\code{\link{lotek_num_sats}}
}
\keyword{internal}
