% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotek_auth.R
\name{lotek_token}
\alias{lotek_token}
\title{Format Access Token as String}
\usage{
lotek_token()
}
\value{
Character access token for API calls.
}
\description{
Format Access Token as String
}
\keyword{internal}
