% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Create an interactive html map of animal locations.}
\usage{
make_map(x)
}
\arguments{
\item{x}{A dataframe containing collar data with the columns "id", "lat", "lon" and  "date_time".}
}
\value{
A leaflet map.
}
\description{
Create an interactive html map of animal locations.
}
\examples{
\dontrun{
  collardata \%>\%
  filter_last_loc() \%>\%
  make_map()
}
}
