% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_gpx.R
\name{produce_gpx}
\alias{produce_gpx}
\title{Save collar locations to disk as a GPX file}
\usage{
produce_gpx(x, lat = "lat", lon = "lon", dir = here::here(),
  file_name = paste0("GPS Locations ", format(Sys.time(),
  "\%Y-\%m-\%d \%H\%M\%S"), ".gpx"), crs = 4326, ...)
}
\arguments{
\item{x}{A dataframe containing collar data with column "date_time" that can be coerced into class "Date"}

\item{lat}{A character string that specifies the name of latitude column in the supplied dataframe.}

\item{lon}{A character string that specifies the name of longitude column in the supplied dataframe.}

\item{dir}{The directory to save the GPX file output. Default is the current working directory.}

\item{file_name}{The name of the GPX file created with this function.}

\item{crs}{The coordiante refernce system associated with the collar data in the supplied data frame.}

\item{...}{Additional arguments to pass to sf::st_write()}
}
\value{
The original data frame passed to the function.
}
\description{
Save collar locations to disk as a GPX file
}
\examples{
\dontrun{
  produce_gpx(collar_data, start = Sys.date() - 30)
}
}
