% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_map.R
\name{save_map}
\alias{save_map}
\title{Save interactive maps to html file}
\usage{
save_map(x, file = "./map.html")
}
\arguments{
\item{x}{A leaflet class object e.g. as created with make_map.}

\item{file}{The file path where the output html file should be saved.}
}
\value{
The original object passed to the funciton unmodified.
}
\description{
Save interactive maps to html file
}
\examples{
\dontrun{
  collar_data \%>\%
  collar::make_map() \%>\%
  collar::save_map()
}
}
