% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_scale_color.R
\name{style_scale_color}
\alias{style_scale_color}
\title{Color Scale Constructor for ggplot}
\usage{
style_scale_color(palette = "main", discrete = T, reverse = F, ...)
}
\arguments{
\item{palette}{Character name of palette in cllr_palettes}

\item{discrete}{Boolean; Is the color aesthetic discrete or not?}

\item{reverse}{Boolean; Should the colors be reversed?}

\item{...}{Additional arguments which can be passed to ggplot2::disrete_scale()
or ggplot2::scale_color_gradientn(), used when discrete is
true or false, respectively}
}
\value{
An object of class "ggproto", for use in ggplot2::ggplot
}
\description{
The purpose of style_scale_color is to make our own colors!
             This includes the "duck hunt" color palette!
}
\details{
There are four color schemes to choose from:
         1) "main" = simple colors, duckhunt; "NULL" value
         2) "duckhunt" = palette from duckhunt
         3) "cblind" = color blind palette
         4) "gray" = gray scale if you're boring
         If discrete = F, will pick gradient between discrete colors
}
