% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_theme.R
\name{style_theme}
\alias{style_theme}
\title{Theme Layout for ggplot}
\usage{
style_theme()
}
\description{
Create a simple ggplot theme for graphs
}
\examples{
# Create a scatterplot with style!
ggplot2::ggplot(iris,
  ggplot2::aes(Sepal.Width, Sepal.Length, color = Species)) +
  ggplot2::geom_point(size = 5) +
  style_theme()
}
