% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trktg_auth.R
\name{trktg_login}
\alias{trktg_login}
\title{Authenticate to Track Tag website}
\usage{
trktg_login(usr, pwd)
}
\arguments{
\item{usr}{username The username associated with the account}

\item{pwd}{password The password used with the supplied username}
}
\value{
True if login succeeds, false if not.
}
\description{
Send username and password info to Track Tag website
  to log in to a user account.
}
\examples{
\dontrun{
trktg_login("WyoDemo", "WYODEMO2024")

fixes <- fetch_trktg_positions()

trktg_logout()
}
}
\seealso{
\code{\link{trktg_logout}} for closing the session.
}
