% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trktg_auth.R
\name{trktg_logout}
\alias{trktg_logout}
\title{Close Track Tag Session}
\usage{
trktg_logout()
}
\value{
True if login info is erased, false if not.
}
\description{
Removes authentication info from memory and logs out of
  Track Tag website.
}
\examples{

trktg_login("some_user", "some_users_pw")

fixes <- fetch_trktg_positions()

trktg_logout()

}
\seealso{
\code{\link{trktg_login}} for starting the session.
}
