% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trktg_auth.R
\name{trktg_token}
\alias{trktg_token}
\title{Track Tag Request Verification Token}
\usage{
trktg_token()
}
\value{
Character request verification token for http requests.
}
\description{
Track Tag Request Verification Token
}
\section{Notes}{


  There is also a cookie called \code{__RequestVerificationToken} set when
  logging in, but it doesn't appear to be used in subsequent requests. The
  token returned here is generated in a hidden input in the login page, and
  must be passed as form data in subsequent requests.
}

\keyword{internal}
