% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vectronics.R
\name{call_vec_api}
\alias{call_vec_api}
\title{A basic call to the Vectronics API}
\usage{
call_vec_api(url, rename_fun = adj_col_nms)
}
\arguments{
\item{url}{a character string representing the url to query}

\item{rename_fun}{Column renaming function}
}
\value{
tibble
}
\description{
A basic call to the Vectronics API
}
\keyword{internal}
