% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllr_dl_vectronics.R
\name{call_vec_api}
\alias{call_vec_api}
\title{A basic call to the Vectronics API}
\usage{
call_vec_api(url)
}
\arguments{
\item{url}{a character string representing the url to query}
}
\value{
A tibble...eventually
}
\description{
A basic call to the Vectronics API
}
\details{
At the time of this writing the API does not return any data.
}
\examples{
#  Build url from base url, collar id, collar key and data type
url <- build_vec_urls(
  base_url = NULL,
  collar_id = get_id_from_key(
    system.file("extdata", package = "CollarDownloadeR")
  ),
  collar_key = get_keys(
    system.file(
      "extdata",
      package = "CollarDownloadeR"
    )
  ),
  type = "act"
)
# Call API - This will not work without a valid key and collar id
# call_vec_api(url)
}
