% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_scale_color.R, R/style_scale_fill.R
\name{cllr_colors}
\alias{cllr_colors}
\title{Grab any combination of colors for our palettes}
\usage{
cllr_colors(...)

cllr_colors(...)
}
\arguments{
\item{...}{colors ya want}

\item{...}{colors ya want}
}
\description{
Grab any combination of colors for our palettes

Grab any combination of colors for our palettes
}
