% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_scale_color.R, R/style_scale_fill.R
\name{cllr_pal}
\alias{cllr_pal}
\title{Access Palettes}
\usage{
cllr_pal(palette = "main", reverse = F, ...)

cllr_pal(palette = "main", reverse = F, ...)
}
\arguments{
\item{palette}{Character name of palette in cllr_palettes}

\item{reverse}{Boolean; Should the colors be reversed?}

\item{palette}{Character name of palette in cllr_palettes}

\item{reverse}{Boolean; Should the colors be reversed?}
}
\description{
Access any of the color palettes available

Access any of the color palettes available
}
