% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_scale_color.R
\docType{data}
\name{cllr_palettes}
\alias{cllr_palettes}
\title{Create palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
cllr_palettes
}
\description{
Color palettes stored in a list
}
\keyword{datasets}
