% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllr_remove_header.R
\name{cllr_remove_header}
\alias{cllr_remove_header}
\title{A function to find and remove data headers form collar data based on the latitude column}
\usage{
cllr_remove_header(x, col_nm, rm_header = TRUE,
  rename_fun = adj_col_nms)
}
\arguments{
\item{x}{A dataframe containing collar data}

\item{col_nm}{The unqouted name of any one column in the data frame}

\item{rm_header}{Logical indicating whether to remove header completely, defaults to TRUE}

\item{rename_fun}{A function used to rename columns, defaults to adj_col_nms, see details}
}
\value{
tibble
}
\description{
A function to find and remove data headers form collar data based on the latitude column
}
\details{
The argument rm_header is meant to aid in programming scenarios where some files should be left alone and others need the header removed. We are not sure this is actually useful, but thought of the scenario and accommodated it just the same.

Column names are adjusted using a custom function, but the user can pass any function they want to manipulate column names (e.g. make.names). The default removes non-ASCII characters, coerces all characters to lower case and replaces "." with "_".
}
\examples{
ugly_file <- read.csv(
  system.file("extdata", "telonics.csv", package = "collar")
)

# The column name needs to be unquoted (Column, not "Column")
# If spaces exist in the column name you may need to use backticks, top left
#  on most(?) qwerty keyboards
collar::cllr_remove_header(ugly_file, `GPS Latitude`, rm_header = TRUE)
}
