% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_last_loc.R
\name{filter_last_loc}
\alias{filter_last_loc}
\title{Find the last location for each individual}
\usage{
filter_last_loc(x, id = "id")
}
\arguments{
\item{x}{A dataframe containing collar data.}

\item{id}{A character vector corresponding the the column to be used as a grouping variable.}
}
\value{
A dataframe with the last location for each individual.
}
\description{
Find the last location for each individual
}
\examples{
\dontrun{
  filter_last_loc(collar_data)
}
}
