% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vectronics.R
\name{get_id_from_fnames}
\alias{get_id_from_fnames}
\alias{get_id_from_key}
\title{Extract collar ID from key}
\usage{
get_id_from_fnames(key_dir, ...)

get_id_from_key(key_dir)
}
\arguments{
\item{key_dir}{path to a single directory containing, potentially many, key files}

\item{...}{Other arguments to pass to key_paths}
}
\value{
character vector of collar IDs
}
\description{
Extract collar ID from key
}
\details{
Vectronics includes the collar ID in the key file name as well as within the XML of the .keyx file.  These functions give the user the option of either extracting the information from the name of the file(s) get_id_from_fnames or reading the ID encoded in the key file get_id__from_key. Storing information in file names or directory names is not a good practice and we would rather have all information stored in files and so prefer and suggest reading the XML within the files.
}
\examples{
collar:::get_id_from_fnames(
  system.file("extdata", package = "collar")
)
collar:::get_id_from_key(
  system.file("extdata", package = "collar")
)
}
\keyword{internal}
