% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllr_dl_vectronics.R
\name{get_key}
\alias{get_key}
\alias{get_keys}
\title{Retrieve keys from a path to a key file or read all keys contained within a directory}
\usage{
get_key(key_path)

get_keys(key_dir)
}
\arguments{
\item{key_path}{A length 1 character string describing the full path to a single key file.}

\item{key_dir}{A length 1 character string describing the full path to a directory containing, potentially many, key files}
}
\value{
A character vector of keys
}
\description{
The key is need to call the Vectronics API and given the length of the key copy and paste errors seems likely.  The functions here help users access the keys in a programmatic fashion.
}
\details{
get_keys is a wrapper for get_key where the former is retained to allow users to access a single file if desired.
}
\examples{
get_key(
  system.file(
    "extdata",
    "Collar123456_Registration.keyx",
    package = "CollarDownloadeR"
  )
)

get_keys(
  system.file(
    "extdata",
    package = "CollarDownloadeR"
  )
)
}
