% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotek_auth.R
\name{lotek_login}
\alias{lotek_login}
\title{Authenticate to Lotek API}
\usage{
lotek_login(usr, pwd, user, pw)
}
\arguments{
\item{usr}{username The username associated with the account}

\item{pwd}{password The password used with the supplied username}

\item{user}{deprecated to align with other functions in collar}

\item{pw}{deprecated to align with other functions in collar}
}
\value{
True if login succeeds, false if not.
}
\description{
Send username and password info to Lotek API
  to log in to a user account.
}
\examples{
\dontrun{
lotek_login("demo", "PASSWORD09")

alerts <- fetch_lotek_alerts()

lotek_logout()
}
}
\seealso{
\code{\link{lotek_logout}} for closing the session.
}
