% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotek_auth.R
\name{lotek_refresh_token}
\alias{lotek_refresh_token}
\title{Refresh Lotek Access Token}
\usage{
lotek_refresh_token(force_refresh = FALSE)
}
\arguments{
\item{force_refresh}{Logical - only used for testing.}
}
\value{
True if current token is still valid or token is refreshed,
  false if token is expired and attempt to refresh failed.
}
\description{
Attempts to acquire a new access token when current one expires.
}
\section{Notes}{


  This function only works sporadically and the reasons are unclear.
  If this function fails a new access token can be created by resending
  the user name and password with lotek_login.
}

\seealso{
\code{\link{lotek_login}} for starting a new session.
}
\keyword{internal}
