% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gpx.R
\name{make_gpx}
\alias{make_gpx}
\title{Save collar locations to disk as a GPX file}
\usage{
make_gpx(x, file, lon_col = lon, lat_col = lat, dt_col, name_col, desc_col)
}
\arguments{
\item{x}{a data frame containing collar data with latitude and longitude location data}

\item{file}{the file path where the output GPX file should be saved}

\item{lon_col}{unquoted name of column containing longitude coordinates}

\item{lat_col}{unquoted name of column containing longitude coordinates}

\item{dt_col}{unquoted name of column containing date and time of location. Date and time in are in Universal Coordinated Time (UTC) and must be formatted as %y-%m-%dT%H-%M-%SZ.}

\item{name_col}{unquoted name of column containing the name to be assigned to the GPS location}

\item{desc_col}{unquoted name of column containing a a description to be assigned to the GPS location}
}
\value{
The original data frame passed to the function.
}
\description{
Save collar locations to disk as a GPX file
}
\examples{
\dontrun{
  make_gpx(collar_data)
}
}
