% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gpx.R
\name{make_gpx}
\alias{make_gpx}
\title{Save collar locations to disk as a GPX file}
\usage{
make_gpx(x, file, lon = lon, lat = lat, date_time, name, desc)
}
\arguments{
\item{x}{a data frame containing collar data with latitude and longitude location data}

\item{file}{the save location and name of the GPX file to be created}

\item{lon}{unquoted name of column containing longitude coordinates}

\item{lat}{unquoted name of column containing longitude coordinates}

\item{date_time}{unquoted name of column containing date and time of location. The format of date_time must follow YYYY-MM-DD HH:MM:SS or YYYY-MM-DDTHH:MM:SS.}

\item{name}{unquoted name of column containing the name to be assigned to the GPS location}

\item{desc}{unquoted name of column containing a a description to be assigned to the GPS location}
}
\value{
The original data frame passed to the function.
}
\description{
Save collar locations to disk as a GPX file
}
\examples{
\dontrun{
  make_gpx(collar_data)
}
}
