% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gpx.R
\name{make_gpx}
\alias{make_gpx}
\title{Save collar locations to disk as a GPX file}
\usage{
make_gpx(x, id_col = "id", lat_col = "lat", lon_col = "lon",
  dt_col = "dt", file = paste0("./GPS Locations ", format(Sys.time(),
  "\%Y-\%m-\%d \%H\%M\%S"), ".gpx"), crs = 4326, ...)
}
\arguments{
\item{x}{A data frame containing collar data with column "dt" that can be coerced into class "Date"}

\item{id_col}{Quoted name of the ID column that signals the individual identifier.}

\item{lat_col}{Quoted name of the latitude column.}

\item{lon_col}{Quoted name of the longitude column.}

\item{dt_col}{Quoted name of the date/time column with class POSIXct.}

\item{file}{The save location and name of the GPX file created.}

\item{crs}{The coordinate reference system associated with the collar data in the supplied data frame.}

\item{...}{Additional arguments to pass to sf::st_write()}
}
\value{
The original data frame passed to the function.
}
\description{
Save collar locations to disk as a GPX file
}
\examples{
\dontrun{
  make_gpx(collar_data, start = Sys.date() - 30)
}
}
