% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Create an interactive html map of animal locations.}
\usage{
make_map(x, lon_col = lon, lat_col = lat, id_col, dt_col)
}
\arguments{
\item{x}{A data frame containing collar data with latitude and longitude columns}

\item{lon_col}{name of the longitude column}

\item{lat_col}{name of the latitude column}

\item{id_col}{name of the ID column that signals the individual identifier}

\item{dt_col}{name of the date/time column}
}
\value{
A leaflet map.
}
\description{
Create an interactive html map of animal locations.
}
\examples{
\dontrun{
  collardata \%>\%
  filter_last_loc() \%>\%
  make_map()
}
}
