% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_theme.R
\name{style_theme}
\alias{style_theme}
\title{Theme Layout for ggplot}
\usage{
style_theme()
}
\description{
Create a simple ggplot theme for graphs
}
\details{
We can change this later to accomodate opinions...
}
